--- src/components/feed/index.js
@@ -1,4 +1,0 @@
-export { default as Board } from "./Board";
-export { default as Header } from "./Header";
-export { default as Nav } from "./Nav";
-export { default as SearchBar } from "./SearchBar";
