+++ src/components/home/feed/searchBar/SearchBarAnimal.js
