+++ src/components/home/index.js
@@ -1,0 +1,4 @@
+export { default as Board } from "./feed/Board";
+export { default as Header } from "./feed/Header";
+export { default as Nav } from "./feed/Nav";
+export { default as SearchBar } from "./feed/searchBar/SearchBar";
