+++ src/components/index.js
@@ -1,3 +1,3 @@
 export * from "./common";
-export * from "./feed";
+export * from "./home";
 export * from "./join";
